Enable Curses Half-Delay to Reduce CPU Usage
When Curses is in non-blocking mode, getch will immediately return
which is not ideal since the outer loop will iterate and consume much of
the CPU usage. When in half-delay mode, getch will block for a certain
amount of time and then release if no input was received from the
user.