Tweaked Tetrimino Generation Logic
Tetrimino generation was completely randomized, and this made the game
quite difficult to play--even at slow speeds--because you're not
guarenteed to get the piece you need in a reasonable amount of time.

Most implementations of Tetris generate all seven tetriminos and then
randomize when they are used, then regenerate when all tetriminos are
used. This revision adds this functionality.